pyproject.toml
[tool.poetry]
name = "async-image-prediction"
version = "0.1.0"
description = "asynchronous image prediction system. The system will predict whether or not an image has a dog in it or not in an asynchronous matter"
authors = ["Nishanth pradeep raj R <nishanthrajamani@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.6"
uvicorn = "^0.34.0"
pydantic = "^2.10.4"
celery = "^5.4.0"
redis = "^5.2.1"
sqlalchemy = "^2.0.36"
torch = "^2.5.1"
torchvision = "^0.20.1"
ultralytics = "^8.3.55"
python-dotenv = "^1.0.1"
loguru = "^0.7.3"
python-multipart = "^0.0.20"
psycopg2-binary = "^2.9.10"
flower = "^2.0.1"
yolo5 = "^0.0.1"
huggingface-hub = "0.10.0"
httpx = "^0.28.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-cov = "^6.0.0"
black = "^24.10.0"
flake8 = "^7.1.1"
pre-commit = "^4.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
